Refactor PostService structure and update related utilities
Refactored PostService by incorporating PostServiceUtils for streamlined validation and retrieval logic, replacing redundant methods like isPostExists with getPost. Updated dependencies across services, tests, and controllers to adapt to these changes. Enhanced data validation by introducing DataValidationException for better error handling.
