+++ src/test/java/faang/school/postservice/service/post/PostServiceTest.java
@@ -5,8 +5,7 @@ import faang.school.postservice.dto.post.PostDto;
 import faang.school.postservice.mapper.post.PostMapperImpl;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.PostRepository;
-import faang.school.postservice.service.project.ProjectService;
-import faang.school.postservice.service.user.UserService;
+import faang.school.postservice.service.PostService;
 import faang.school.postservice.service.utils.PostServiceUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -76,7 +75,7 @@ class PostServiceTest {
         @Test
         void testPublishPostFalseFromStart() {
             testPost.setPublished(false);
-            when(postServiceUtils.isPostExists(anyLong())).thenReturn(testPost);
+            when(postServiceUtils.getPost(anyLong())).thenReturn(testPost);
             when(postRepository.save(testPost)).thenReturn(testPost);
 
             PostDto postDto = postService.publishPost(1L);
@@ -89,7 +88,7 @@ class PostServiceTest {
     @Test
     void testPublishPostTrueFromStart() {
             testPost.setPublished(true);
-        when(postServiceUtils.isPostExists(1L)).thenReturn(testPost);
+        when(postServiceUtils.getPost(1L)).thenReturn(testPost);
         assertThrows(IllegalArgumentException.class,
                 () -> postService.publishPost(1L));
         }
@@ -97,11 +96,12 @@ class PostServiceTest {
         @Test
         void testUpdate() {
             String content = "Updated content";
-            when(postServiceUtils.isPostExists(anyLong())).thenReturn(testPost);
+            when(postServiceUtils.getPost(anyLong())).thenReturn(testPost);
 
             postService.update(1L, content);
             assertEquals(content, testPost.getContent());
         }
+
     }
